+++ src/main.py
@@ -1,5 +1,4 @@
-from controller import main
+from controller.controller import main
 main()
 
-#Crea tablas manualmente, automatizar; TODO generar modulo que cree toda la DB y meterlo en el controlador como primer paso al correr la app# historia_academica
 #TODO validar ints, strings y floats# Todo rowGetter
