+++ src/modules/materia.py
@@ -1,19 +1,6 @@
-#TODO all module
-from datetime import date
 import sqlite3 as sql
 from decouple import config
-
 DB = config('DB_NAME')
-
-#Crea base de datos, hay que hacer que verifique si ya existe
-def createrDB():
-    try:
-        conn = sql.connect(DB);
-        conn.commit();
-        conn.close();
-    except sql.Error as e:
-        print(e)
-
 
 #Crea tablas manualmente, automatizar
 def createTeable():
@@ -21,51 +8,44 @@ def createTeable():
         conn = sql.connect(DB)
         cursor = conn.cursor()
         cursor.execute(
-            """CREATE TABLE IF NOT EXISTS estudiante (
-                identificacion INTEGER PRIMARY KEY,
+            """CREATE TABLE IF NOT EXISTS materias (
+                codigo INTEGER PRIMARY KEY,
                 nombre TEXT NOT NULL,
-                apellido TEXT NOT NULL,
-                carrera TEXT NOT NULL,
-                fechanacimiento TEXT NOT NULL,
-                fechaingreso TEXT NOT NULL,
-                procedencia TEXT NOT NULL,
-                correoeletronico TEXT NOT NULL,
-                cantidadmatriculas INTEGER NOT NULL
+                facultad TEXT NOT NULL,
+                departamento TEXT NOT NULL,
+                idioma TEXT NOT NULL,
+                creditos INTEGER NOT NULL
                 )"""
         );
         conn.commit();
         conn.close();
     except sql.Error as e:
         print(e)
-
-
-        
+    
 #Inserta la materia
-def insertRow(identificacion : int, nombre: str, apellido: str, carrera: str, fechanacimiento: str, fechaingreso: str, procedencia: str, correoeletronico: str, cantidadmatriculas: int):
+def insertRow(codigo: int, nombre: str, facultad: str, departamento: str, idioma: str, creditos: str):
     try:
         conn = sql.connect(DB)
         cursor = conn.cursor()
-        instruction = f"INSERT INTO estudiante values({identificacion}, '{nombre}', '{apellido}', '{carrera}', '{fechanacimiento}', '{fechaingreso}','{procedencia}','{correoeletronico}','{cantidadmatriculas}')"
+        instruction = f"INSERT INTO materias values({codigo}, '{nombre}', '{facultad}', '{departamento}', '{idioma}', {creditos})"
         cursor.execute(instruction)
         conn.commit();
         conn.close();
     except sql.Error as e:
         print (e)
 
+    
 #Pide input por teclado a tráves de consola de los datos, en versión gráfica desaparece
+#TODO validar ints, strings y floats
 def rowGetter():
-    #TODO validar ints, strings y floats
-    identificacion = input('numero de identificacion del estudiante: ')
-    identificacion = identificacion.ljust(10)
-    nombre = input('Nombre del estudiante: ')
-    apellido = input('apellido del estudiante: ')
-    carrera = input('nomrbre de la carrera: ')
-    fechanacimiento = input('fecha de nacimiento del estudiante: ')
-    fechaingreso = input('fecha de ingreso del estudiante: ')
-    procedencia = input('procedencia del estudiante: ')
-    correoeletronico = input('corre oeletronico del estudiante: ')
-    cantidadmatriculas = input('cantidad de matriculas del estudiante: ')
-    return (int(identificacion), nombre.upper(), apellido.upper(), carrera.upper(), fechanacimiento.upper(), fechaingreso.upper(), procedencia.upper(),correoeletronico.upper(),int(cantidadmatriculas))
+    codigo = input('Codigo de la materia: ')
+    codigo = codigo.ljust(10)
+    nombre = input('Nombre de la materia: ')
+    facultad = input('Facultad que dicta la materia: ')
+    departamento = input('Departamento que dicta la materia: ')
+    creditos = input('Creditos de la materia: ')
+    idioma = input('Idioma en que se dicta la materia: ')
+    return (int(codigo), nombre.upper(), facultad.upper(), departamento.upper(), idioma.upper(), int(creditos))
 
 #pide varias veces los datos
 def batchRowGetter():
@@ -73,17 +53,14 @@ def batchRowGetter():
     secret_runner = "1"
     counter = 0
     while True: 
-        identificacion = input('numero de identificacion del estudiante: ')
-        identificacion = identificacion.ljust(10)
-        nombre = input('Nombre del estudiante: ')
-        apellido = input('apellido del estudiante: ')
-        carrera = input('nomrbre de la carrera: ')
-        fechanacimiento = input('fecha de nacimiento del estudiante: ')
-        fechaingreso = input('fecha de ingreso del estudiante: ')
-        procedencia = input('procedencia del estudiante: ')
-        correoeletronico = input('corre oeletronico del estudiante: ')
-        cantidadmatriculas = input('cantidad de matriculas del estudiante: ')
-        result.append((int(identificacion), nombre, apellido, carrera, fechanacimiento, fechaingreso, procedencia, correoeletronico, int(cantidadmatriculas)))
+        codigo = input('Codigo de la materia: ')
+        codigo = codigo.ljust(10)
+        nombre = input('Nombre de la materia: ')
+        facultad = input('Facultad que dicta la materia: ')
+        departamento = input('Departamento que dicta la materia: ')
+        creditos = input('Creditos de la materia: ')
+        idioma = input('Idioma en que se dicta la materia: ')
+        result.append((int(codigo), nombre, facultad, departamento, idioma, int(creditos)))
         runner=input('Digite 1 si desea continuar, digite cualquier otra tecla si no. ')
         counter+=1
         if runner != secret_runner:
@@ -91,12 +68,13 @@ def batchRowGetter():
     print (f"Usted ha insertado {counter} datos, los cuales son: {result}")
     return result
 
+
 #Leer todos datos
 def readAllRows():
     try:
         conn = sql.connect(DB)
         cursor = conn.cursor()
-        instruction = f"SELECT * from estudiante"
+        instruction = f"SELECT * from materias"
         cursor.execute(instruction)
         datos = cursor.fetchall()
         conn.commit();
@@ -105,57 +83,58 @@ def readAllRows():
     except sql.Error as e:
         print (e)
 
- #Acá se puede filtrar los datos bajo una condición 
+#Acá se puede filtrar los datos bajo una condición 
 def searchByFilter(fieldName, fieldValue):
     try:
         conn = sql.connect(DB)
         cursor = conn.cursor()
-        instruction = f"SELECT * from estudiante WHERE {fieldName}={fieldValue}"
+        instruction = f"SELECT * from materias WHERE {fieldName}={fieldValue}"
         cursor.execute(instruction)
         datos = cursor.fetchall()
         conn.commit();
         conn.close()
         print(datos);
     except sql.Error as e:
-        print (e)   
+        print (e)
 
 #Batch write, escribe multiples lineas de datos a la vez
 def batchInsertRow(dataList):
     try:
         conn = sql.connect(DB)
         cursor = conn.cursor()
-        instruction = f"INSERT INTO estudiante values(?, ?, ?, ?, ?, ?, ?, ?, ?)"
+        instruction = f"INSERT INTO materias values(?, ?, ?, ?, ?, ?)"
         cursor.executemany(instruction, dataList)
         conn.commit();
         conn.close();
     except sql.Error as e:
-        print (e)   
+        print (e)
+    
 
-#Read order, ordena los datos de mayor a menor según el campo que le pidamos 
+#Read order, ordena los datos de mayor a menor según el campo que le pidamos :p
 def readOrdered(field: str):
     try:
         conn = sql.connect(DB)
         cursor = conn.cursor()
-        instruction = f"SELECT * from estudiante ORDER BY '{field}' DESC" #Si le quitamos el DESC se ordenara de menor a mayor, "DESC" viene de DESCENDING
+        instruction = f"SELECT * from materias ORDER BY {field} DESC" #Si le quitamos el DESC se ordenara de menor a mayor, "DESC" viene de DESCENDING
         cursor.execute(instruction)
         datos = cursor.fetchall()
         conn.commit();
         conn.close()
         print(datos);
     except sql.Error as e:
-        print (e) 
+        print (e)
 
-#Actualizar datos en un determinado campo de algun estudiante
+#Actualizar datos en un determinado campo de alguna materia
 """Actualizar materia en diseño lógico."""
-def update(fieldOnChange: str, dataOnChange, iden: int):
+def update(fieldOnChange: str, dataOnChange, code: int):
     try:
         conn = sql.connect(DB)
         cursor = conn.cursor()
         try:
             dataOnChange= int(dataOnChange)
-            instruction = f"UPDATE estudiante SET '{fieldOnChange}'={dataOnChange} WHERE identificacion={iden}"
+            instruction = f"UPDATE materias SET '{fieldOnChange}'={dataOnChange} WHERE codigo={code}"
         except ValueError:
-            instruction = f"UPDATE estudiante SET '{fieldOnChange}'='{dataOnChange}' WHERE identificacion={iden}" #Comando de actualización en SQL
+            instruction = f"UPDATE materias SET '{fieldOnChange}'='{dataOnChange}' WHERE codigo={code}" #Comando de actualización en SQL
         finally:
             cursor.execute(instruction)
             conn.commit();
@@ -164,13 +143,10 @@ def update(fieldOnChange: str, dataOnChange, iden: int):
     except sql.Error as e:
         print (e)
 
+#---------------------------------------------------------------Función principal---------------------------------------------------------------------------------#
 def main():
     
     while True:
-        #Verifica si existe el archivo.db SOTISFICAR TODO
-        createrDB()
-        # Verifica si existe la tabla    SOTISFICAR TODO
-        createTeable()
         #Revisa si va a añadir datos, leer o actualizar, metodo reutilizable, verifica que el input sea correcto
         validator = True
         while validator:
@@ -184,12 +160,12 @@ def main():
             except ValueError:
                 print("Input invalido")
                 validator = True
-        #-------------------------------------------Creación de estudiante---------------------------------------------------------------------------#
+        #-------------------------------------------Creación de materias---------------------------------------------------------------------------#
         if selector ==1:
             validator = True
             while validator:
                 try:
-                    pointer = input("Si desea solo añadir un estudiante digite '1' y luego enter, si desea registrar multiples estudiantes digite '2' y luego enter. ")
+                    pointer = input("Si desea solo añadir una materia digite '1' y luego enter, si desea registrar multiples datos digite '2' y luego enter. ")
                     pointer = int(pointer)
                     validator = False
                     while(pointer!=1 and pointer !=2):
@@ -201,7 +177,7 @@ def main():
         #Primer caso del input de escritura, un solo dato
             if int(pointer)==1:
                 data = rowGetter()
-                insertRow(data[0], data[1], data[2], data[3], data[4], data[5], data[6], data[7], data[8])
+                insertRow(data[0], data[1], data[2], data[3], data[4], data[5])
                 print("Datos insertados: ",data)
         #Segundo caso, múltiples datos
             elif(int(pointer)==2):
@@ -214,30 +190,27 @@ def main():
             validator = True
             while validator:
                 try:
-                    iden = input("Escriba el numero de identificacion del estudiante que quiere actualizar ")
-                    iden = int(iden)
+                    code = input("Escriba el codigo que quiere actualizar ")
+                    code = int(code)
                     validator = False
                 except ValueError:
                     print("Input invalido")
                     validator = True
             field = input("""Escriba por el campo que quiere actualizar, recuerde que los campos son
-                    identificacion,
+                    codigo,
                     nombre
-                    apellido,
-                    carrera,
-                    fechanacimiento,
-                    fechaingreso
-                    procedencia
-                    correoeletronico
-                    cantidadmatriculas
+                    facultad,
+                    departamento,
+                    idioma,
+                    creditos
             """)
-            dataOnchange = input(f"Escriba el valor con el cual quiere modificar el campo {field.lower()} de la materia con identificacion{iden} ")
+            dataOnchange = input(f"Escriba el valor con el cual quiere modificar el campo {field.lower()} de la materia con codigo{code} ")
             try:
                 dataOnchange = int(dataOnchange)
             except:
                 dataOnchange.lower()
             finally:
-                update(field, dataOnchange, iden)
+                update(field, dataOnchange, code)
 
         #------------------------------------------------------------------------------------------------------------------------------------------#
 
@@ -251,13 +224,10 @@ def main():
                 field = input("""Escriba por el campo que quiere filtrar, recuerde que los campos son
                     codigo,
                     nombre
-                    apellido,
-                    carrera,
-                    fechanacimiento,
-                    fechaingreso
-                    procedencia
-                    correoeletronico
-                    cantidadmatriculas
+                    facultad,
+                    departamento,
+                    idioma,
+                    creditos
                 """)
                 readOrdered(field.lower())
             elif order==2:
@@ -265,5 +235,3 @@ def main():
         #-----------------------------------------------------------------------------------------------------------------------------------------#
         elif selector ==4:
             break;
-
-#-----------------------------------------------------------------------------------------------------------------------------------------------------------------#
